Another pass.
reflekt is 419,802

boon-json is 115,315

boon-util is 93,718

boon 0.32 was 1.1 MB

To use boon-json which is the parser, you only need reflekt which is more or less boon-core and boon-json.

So we went from 1.1 MB to 550K so about 1/2 the size of boon 0.32.

This is removing a lot of non-essential stuff like data-repo (which will be another project not in boon).